Adding metadata to dataset
Went back through the dataset, and added information about which paragraph, or which figure contains information that we decided was good enough to code a given procedural unit as present or absent.

Also, made sure that the data were consistent with a newer version of the codebook, 1.1, which is currently in the development branch of the procedural unit codebook github repo. Main change to that codebook is just adding more information about inclusion/exclusion criteria, and removal of bifacial retouch as a procedural unit, which did not make an impact on the analysis.